+++ src/EthernaCredit.Domain/Models/OperationLogs/WithdrawOperationLog.cs
@@ -4,7 +4,7 @@
     {
         // Constructors.
         public WithdrawOperationLog(
-            double ammount,
+            decimal ammount,
             string author,
             User user)
             : base(ammount, author, user)
